+++ cc-lib/miniz.cc
@@ -2440,12 +2440,12 @@ tinfl_status tinfl_decompress(tinfl_decompressor *r, const mz_uint8 *pIn_buf_nex
     if (((out_buf_size_mask + 1) & out_buf_size_mask) || (pOut_buf_next < pOut_buf_start))
     {
       if (DEBUG_MINIZ) {
-        printf("%p - %p + %lld - 1\n",
-               pOut_buf_next, pOut_buf_start, (int64_t)*pOut_buf_size);
-        printf("%lld + %lld - 1\n", (int64_t)(pOut_buf_next - pOut_buf_start),
+        printf("%p - %p + %zu - 1\n",
+               pOut_buf_next, pOut_buf_start, (size_t)*pOut_buf_size);
+        printf("%zu + %zu - 1\n", (size_t)(pOut_buf_next - pOut_buf_start),
                (int64_t)*pOut_buf_size);
-        printf("%lld, %p < %p\n",
-               (int64_t)out_buf_size_mask, pOut_buf_next, pOut_buf_start);
+        printf("%zu, %p < %p\n",
+               (size_t)out_buf_size_mask, pOut_buf_next, pOut_buf_start);
       }
         *pIn_buf_size = *pOut_buf_size = 0;
         return TINFL_STATUS_BAD_PARAM;
